+++ exe/quke_demo_app
@@ -38,6 +38,6 @@ if ENV["COVERAGE"]
   SimpleCov.start
 end
 
-require_relative "../lib/quke/demo_app"
+require_relative "../lib/quke_demo_app"
 
-Quke::DemoApp::Cli.start
+QukeDemoApp::Cli.start
