--- lib/quke.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-require "quke/demo_app"
-
-module Quke
-
-end
