--- lib/quke/demo_app.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "demo_app/concerns/can_have_search_results"
-require_relative "demo_app/app"
-require_relative "demo_app/cli"
-
-module Quke
-  module DemoApp
-    # The DemoApp namespace
-  end
-end
