--- lib/quke/demo_app/version.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Quke
-  module DemoApp
-    VERSION = "0.1.1"
-  end
-end
