+++ lib/quke_demo_app.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require_relative "quke_demo_app/concerns/can_have_search_results"
+require_relative "quke_demo_app/app"
+require_relative "quke_demo_app/cli"
+
+module QukeDemoApp
+
+end
