+++ lib/quke_demo_app/public/assets/javascript/ie10-viewport-bug-workaround.js
