+++ lib/quke_demo_app/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module QukeDemoApp
+  VERSION = "0.1.1"
+end
