+++ lib/quke_demo_app/views/_partial_example.erb
