+++ lib/quke_demo_app/views/index.erb
