+++ lib/quke_demo_app/views/javascript_error.erb
