+++ lib/quke_demo_app/views/layout.erb
