+++ lib/quke_demo_app/views/search.erb
