+++ spec/support/quke_demo_app.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 
 # Need to require our actual code files.
-require "quke/demo_app"
+require "quke_demo_app"
